Factor loadResults out of QueryBuilderRoute.js
