+++ src/util/loadResults.js
@@ -1,0 +1,50 @@
+import cloneDeep from 'lodash.clonedeep';
+import { v4 as uuidv4 } from 'uuid';
+import stripesFetch from './stripesFetch';
+
+const loadResults = async (intl, stripes, values, setQueryResponse, setError) => {
+  try {
+    const limit = values.tables[0].limit;
+    const modifiedValues = cloneDeep(values);
+    modifiedValues.tables[0].limit = parseInt(limit, 10) + 1;
+
+    const resp = await stripesFetch(stripes, '/ldp/db/query', {
+      method: 'POST',
+      body: JSON.stringify(modifiedValues),
+    });
+    if (!resp.ok) throw new Error(`XXX HTTP error ${resp.status}: ${resp.statusText}`);
+    resp
+      .json()
+      .then(jsonResp => {
+        jsonResp.forEach(v => { delete v.data; });
+        const isComplete = jsonResp.length < limit;
+
+        if (!isComplete) {
+          let firstField;
+
+          // ### I don't know if this is guaranteed to work, but it seems to
+          Object.keys(jsonResp[0]).forEach(key => {
+            if (!firstField) firstField = key;
+          });
+
+          jsonResp[jsonResp.length - 1] = {
+            [firstField]: '... More records ...',
+          };
+        }
+
+        setQueryResponse({
+          key: uuidv4(),
+          count: isComplete ? jsonResp.length : limit,
+          isComplete,
+          resp: jsonResp,
+        });
+      })
+      .catch((e) => {
+        setError(e.toString());
+      });
+  } catch (err) {
+    setError('XXX Query failed: ' + err);
+  }
+};
+
+export default loadResults;
